requirements.txt
Flask
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas
requests
beautifulsoup4
